ahora le di retoqueal código para agregarle más funcionalidad decimal

+++ Program.cs
@@ -6,7 +6,7 @@ namespace semana_2__proyecto_de_consola_calculadora
     {
         static void Main(string[] args)
         { // Declare variables and then initialize to zero.
-            int num1 = 0; int num2 = 0;
+            int num1 = 0; float num2 = 0;
 
             // Display title as the C# console calculator app.
             Console.WriteLine("Console Calculator in C#\r");
